Move Sourcemap from model to sourcemap handler
Reserve model for event types, not auxiliary documents like this.
