+++ beater/api/asset/sourcemap/sourcemap_test.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package model_test
+package sourcemap
 
 import (
 	"context"
@@ -28,7 +28,6 @@ import (
 
 	"github.com/elastic/beats/v7/libbeat/common"
 
-	"github.com/elastic/apm-server/model"
 	"github.com/elastic/apm-server/transform"
 )
 
@@ -38,7 +37,7 @@ func getStr(data common.MapStr, key string) string {
 }
 
 func TestTransform(t *testing.T) {
-	p := model.Sourcemap{
+	p := sourcemapDoc{
 		ServiceName:    "myService",
 		ServiceVersion: "1.0",
 		BundleFilepath: "/my/path",
@@ -59,7 +58,7 @@ func TestTransform(t *testing.T) {
 }
 
 func TestParseSourcemaps(t *testing.T) {
-	fileBytes, err := ioutil.ReadFile("../testdata/sourcemap/bundle.js.map")
+	fileBytes, err := ioutil.ReadFile("../../../../testdata/sourcemap/bundle.js.map")
 	assert.NoError(t, err)
 	parser, err := s.Parse("", fileBytes)
 	assert.NoError(t, err)
